GetPacket() lengthing nanosleep() ensure we get entire packet.
Not a good long term fix.  The problem is the non-blocking read().
Not sure how to tell whether a read length of ZERO means the rest of
the packet is not in yet, or already done.  When I figure that out I
can break the loop and return to the master select().