Open the shared-memory segment to talk to ntpd only once.
Instead of opening the shared-memory segment to talk to ntpd each time 
gpsd_init is called and having it be part of the GPS device state, open
it once at daemon startup time and move the pointer into the shared-context
structure.

This is a good cleanup on general principles.   It will also let us drop
root privs after startup.