Fix a serious bug in multi-device operation.
Rob Janssen quite correctly reported as a bug the fact that when gpsd
went hunting on a new device it fell out of the main loop and ignored
pending input from devices already connected.  This is because 
packet_sniff was written to loop until it found a packet or timed out.

The fix: opening a device no longer does a packet sniff.  Instead,
gpsd_poll() doubles as the packet sniffer when the driver-type pointer
is null, and the driver-type pointer for the device is only set when
it finds a whole packet.