Rob Janssen writes: Here is my first version of the Trimble TSIP driver.
It works with my Trimble Lassen SK II receiver.

Time and Position look OK, Velocity I can't test.
There are still issues I am trying to solve.  Most notably, the receiver
does not output satellite AZ/EL in any normal packet, only in debug
packets and raw measurement data packets (that are quite large).   So the
constellation map is not working.

As usual, the overall picture is that the protocol is very sub-optimal.
The framing sucks.  The data is sent in very small messages that contain
only one type of info.  The values are in floating point.   Some data can
only be polled, not sent at regular intervals.

However, the receiver itself is quite good, even though it is 8-channel.
Accuracy is good, and it is clear that attention has been paid to the
content of the time message and the exact moment it is sent.

IMHO the quality ranking of the receivers I have is:

1. Trimble
2. Rockwell
3. SiRF.

This seems consistent with the typical situation that the stuff that is
most widely used is inferior in quality :-)