ESR pointed out that I wasn't actually calling serialConfig() anywhere...
which was wrong.  The serialConfig function sets up a number of useful
parameters like flow control, parity checking and VMIN/VTIME.

This only needs to be done once, so we do it right after we open the port.
Other port changes are baud rate only, so we call a different function to
just set the baud rate.