>From Jeff Francis:
I don't generally run X on the car (truck) computer, so I threw
together a quick, dirty, and nasty little C/curses program that does
basically the same thing as xgps, minus the "x" part.  I unashamedly
pillaged xgps.c for some of the code, and unashamedly hard-coded it
for English units (mph and ft).  It's probably not perfect, nor is it
totally bug-free (for example, ^C doesn't work - got that wrong
somewhere).  I don't feel it's up to snuff yet for include/, but I
thought I'd offer it up for use and abuse for anyone who needs a
text-only display.  I'll clean it up as I go (for example, remove the
zillion #ifdefs for curses/non-curses).