Subtle change in how devices and the nowait option are handled.
Before, when a device was opened, gpsd allocated a channel block
for it but didn't read from it unless the nowait (-n) option was on.
No reads would be done from the device until a user session bound it
and requested information.

Now, when a device is opened (either at startup because it's a command-line
argument or because of a hotplug notification) gpsd starts reading
from it immediately in an attempt to lock on to a packet type. Reading
continues until a packet type is identified.  At that point, if there
are no subscribers to the device and the nowait option is off, it
closes the device.

This change trades a few milliseconds more power expenditure at
startup for early knowledge of the device type. Among other things,
this means the I command will work before any user has actually
requested information from the device.  

Soon, this feature will make it possible for clients to select either
GPS or RTCM data to view, because gpsd now knows which kind of devices
are supplying which at the time it fulfills a client request to 
be bound to a channel.

The code changes for this feature are confined to gpsd.c.  It actually
simplifies nowait handling somewhat and reduces the total complexity
of the code.