Bug fixes for xgpsspeed
>From Bill Marr on gpsd-dev
.------------------------------
Attached is a patch against 'xgpsspeed.c' in the 2.30 release of 'gpsd'. It
fixes the following bugs in the 'xgpsspeed' application:

(1) Missing limit-checks on the tachometer widget can result in the tachometer
needle spinning madly if invalid vehicle speeds are reported by 'gpsd'.

(2) The display of speed units (e.g. "Miles per Hour") above the "speedometer"
was erroneously always displaying "Km per Hour", regardless of the speed
units in use. Furthermore, although there was already a provision to select
"knots" as the speed units ("-speedunits knots"), there was never any code to
display the proper text!

(3) An erroneous conversion factor (the reciprocal of the correct value) was
used to convert the native 'meters per second' GPS-reported speed into knots.
This results in an erroneous needle position when "-speedunits knots" has
been specified on the command line.
.------------------------------