Type length meditations.
int is not supposed to be longer than a long int, and a long int is supposed
to be at least 32 bits. Unfortunately that doesn't tell us how many bits are
in a long, so let's use a type where we do know.

I don't suppose anyone is running this on a machine where (we haven't heard
any complaints so far)
	sizeof(int) == 2
	sizeof(int) == 8 && sizeof(long) == 8

RTCM regress is happy on ia32 and amd64.