Add the | command.
This command allows gpsd to pass through a raw command from the client
to the server; the reverse of super-raw (R=2) mode. This could
conceivably be used to reflash your receiver's firmware, modify the
port settings, tune the navigation engine to produce flawed results or
any other thing one could do with a direct connection. This command is
primarily for debugging purposes.