Introducing a new configuration option "--disable-reconfigure"...
...which is used to prevent gpsd from taking actions which would alter
the configuration of your receiver. This could include, but might not
be limited to:

	* port speed
	* communication protocol
	* output messages sent
	* output message rate
	* resetting the receiver
	* clearing stored almanac/ephemeris/position/clock calibration

This is useful in cases where you have your receiver set up in a particular
manner (logging raw pseudoranges, carrier phases and NAV data, for example)
and don't mind sharing that data, so long as it's read-only.

For now, this affects the speed/protocol switch commands in gpsd.c, but the
device drivers will need to be patched as well.