Add a new capability to the generic driver structure: wakeup...
...a function to be called just after the autobaud hunt sets the line
speed each time.  Use this to get rid of the internal baud hunt loop
in the True North driver; instead, the ID query and rate-setting
strings will be sent by the wakeup hook.

Note: this patch is untested.  I'm pretty certain it will work, but
somebody needs to try it on live hardware.