Recognize use of @ as a private packet leader character...
...which both the True North and the Sony CXD2951 do.  It shouldn't be
included in the packet checksum.  This enables us to get rid of the
separate tnt_send() function.

Note: this changes, very slightly, gpsd's behavior with respect to 
Sony CXD2951 chips -- there is an initializer string in drivers.c 
that used to *include* the @ in its checksum.  I suspect this was a
bug, but I'm noting it here in case I just broke the support for that
chip.