OK, I'm slow today.
If we don't already know the type of a device, we have to send it
*every* driver's wakeups on a baud-rate change.  This does, basically,
the same thing the old probe function in the TrueNorth driver did
since at the moment only that driver has a wakeup.

Also, move the send so it gets done when we change baud rates without
closing the device.