Axiom Sandpiper II (SiRF-I) works.
One of the longest-lived bugs bites the dust.  Axiom Sandpiper II is now
tested and working. David Mandala's bug turned out to be a consequence of
two unrelated problems: (1) The Sandpiper doesn't respond to a normal
SiRF probe, and (2) SiRFs echo anything unknown you send at them, so you
get false positives if the positive respone is the same as the probe string.

We're archiving the documentation here because it's only findable one place
on the web (at <http://www.allsurplus.net/Axiom/>), and that location looks 
unstable when the surplus store's supply runs out.