Finish the OpenBSD kernel timestamping hooks.
In recent versions, the kernel keeps timestamps of control line
changes. We also have the nmea(4) line discipline which activates an
in-kernel NMEA parser to extract time information. This is then
exported via the sensors framework for use by openntpd.

This partially worked, but it wasn't until this morning that we had the
control line stuff working.

CK