Fix to a packet-sniffer bug handling $P..
Diego Berge sent me an example of some line noise which just happened to have
$P<garbage> in it. This tricked the packet sniffer into thinking it found an
NMEA device. This change means that now 2 letters are required after the '$'
to trigger a detect. If this still false-positives too much, I might crank it
up to 4, to match on $PFec, $PFST, $PUBX, $PSRF and the like.