Much as I hate to do this, add a command line flag (-b)...
...to prevent gpsd from writing to a device. It sucks that we have to
do this, but I'm really sick of explaining to people that gpsd is not
to blame for the crap being sold as a bluetooth device these days.

This is user-friendly and packager-friendly: users can now run gpsd on an
arbitrary device without fear of wrecking it, and packagers do not need to
produce special "--disable-reconfigure" packages.

In a perfect world
* embedded bluetooth stacks would not be flaky and fragile
* there would be a reliable, cross-platform way to detect a
  bluetooth device, in order to optimize runtime behaviour.

I don't see either of these happening any time soon...