Install gpsd_config.h like I intented to.
Inspired by a diff from "Iain", who is with the gnome org.

Iain's patch also included gpsd_config.h in the headers; I chose
to not include that. I'm not really keen on nested headers. The
patch also included gpsd_config.h in the distribution tarball,
which we don't need to do.