don't return NULL when passed a NULL pointer or a zero-length buffer;
instead, return an empty string. this sort of thing happens when we
call gpsd_hexdump inside one of the packet writers when we are about
to write a packet with no payload. Printing a NULL pointer is rather
ugly... printing an empty string is much prettier and consistent with
the SQUELCH case.