Squish the spin bug.
When many connections are established in a short period of time the
code to detect whether a subscriber slot was in use would allocate the
same slot to multiple clients. "No good can come of this."

To remedy, change the test to check if the fd element indicates
that a subscriber slot is in use. If it seems free, assign a
dummy value. It will be overwritten shortly, but until it is, it
will prevent the slot from being multiply assigned.