Refactoring step and fix to ntp interface.
Extract the code which activates the interface to ntpd
into a separate function and arrange that this is called
by gpsd_assert_sync (in serial.c). This ensures it only
happens once the serial port parameters are known.

Modify the code which probes for devices to issue a
gpsd_assert_sync on success.