Eliminate naive float comparisons.
These were indirectly causing 32-vs-64-bit divergences in the
regression tests -- timestamp comparisons were failing in such a way
that they caused false positives in detecting start-of-cycle on NMEA
devices, leading to fix data being cleared when it should not have
been.  This never affected SiRF chips or other non-NMEA devices, which
is why it took so long to spot.