Backed out some recent changes that broke regression tests.
Specifically, the changes to sirf.c from ckuethe to stop
rAte-linmiting satellite messages (27 Dec 2007) and the changes to
packet.c and packet.states.h from ckuethe on 25 Dec 2007 to make the
packet sniffer 'a bit more picky'.

The packet-getter changes need to be re-implemented in a way
that doesn't break the RTCM regression test.

The SiRF changes will break the regression tests, and that's OK, but
they need to be committed along with a rebuild of the tests so
the tests pass after commit.