Rip out the openbsd-specific timestamping support.
We have a better way in -current: ldattach on the real device, and
relay the data through a pty. ldattach will set up the timestamping
line discipline (there are at least 4 different settings that could be
used) and the relay data out a pty. This is fast - about 60usec. Yay
for the removal of non-portable code that isn't core the purpose of
gpsd. All regression tests pass on openbsd...