rename "FILE * fd" to "FILE * fp"...
I prefer to use names like "fp" for "FILE *" and "fd" for integers
representing a file descriptor.

Fix a small bug related to the first point. The test for failure to open
the log file was basically comparing the return from fopen() against -1.
This is the right test for open(), but fopen() returns NULL on failure.

And clean up some whitespace while I'm here.