Another try at defeating regression-timing bug.
This turns out to be at least a triple-point failure.  Previous
changes reduced the frequency and made it less random, but it appears
gpsd has an unavoidable race condition when feeding through PTYs that
can lose the last input packet.

Workaround: compare test output with both the checkfile and the checkfile with
its last line removed.