sirfmon can read from gpsd again.
The culprit was excessively complicated logic in throttled_write(),
introduced in my throttled_write() change of 12 Jan, that aimed to
make each NMEA packet in a multiple-sentence transmission of faked NMEA
an atomic write.  This was intented to solve the regression-test
timing problem, but didn't.  It turned out to screw up super-raw mode
instead.

Regression tests pass.