Our NMEA field parser was silently ignoring the last (*-terminated) field.
We never noticed this before because, given the sentences we
actually parse, the only piece of data it dropped was VDOP from GPGSA
.-- which we error-model if the GPS doesn't supply it.  So the
practical effect of this fix is to improve the quality of vertical
precision estimation slightly.

Note: found while I was testing gpsmon, so that work has already paid
off.