Python build patches by Richard Hansen ("I" in this comment):
This patch modifies Makefile.am and setup.py so that 'make distcheck'
passes.  This patch makes it possible to do VPATH builds (putting build
products in a separate directory from the unpacked tarball).

Note that I have NOT tested dist-rpm.

The most awkward part of this patch is the changes to setup.py. Python's
distutils package does not have built-in support for VPATH builds, which
significantly complicates things.  Interestingly, Python's own build system
supposedly supports VPATH builds even though it uses distutils.  Python's
setup.py script subclasses several of the distutils packages, so the VPATH
magic must happen somewhere in those hundreds of lines of code.  I'll
investigate further when I have the time.  For now, the changes to setup.py
are fairly minimal and should work.

This patch also causes setup.py to exit with a non-zero return code if the call
to make fails, and adds missing dependencies related to the Python modules.