Even more delayed allocation on wait-for-connect.
Now, we stash each device name given on the command line or presented
over the control port in the channels array, but don't open the device
and sniff packets from it until a user tries to allocate a channel.
At that point we actually need to test each attached device for the
type of data it ships.