Back out recent breakage in the Python binding that didn't work.
Back out bzed's major change of 26 May and three related later changes
by esr on 13 Aug, all in gps.py. These introduced a serious bug - all
pending responses read were unpacked, but *only the last response* in
any given group was available after the return of the poll() call.

This broke gpsfake, a problem which was concealed by a thinko in the
regression-test driver.

The client falling behind the daemon is still an issue, but needs
to be tackled in a different way.