L command disabled...
...because the change from 2 to 3 fields in this command gave Kismet
severe indigestion.

Kismet shouldn't be using this command, but there's no point in
fighting with Dragorn.  We'll just disable it.  The client library
could never parse the response anyway, so the only people who lose are
the ones opening a socket direct to the daemon and doing depreacated
single-shot queries.