J command and switch are gone.
Now that we have reliable end-of-cycle detection in NMEA, we can always 
clear the fixbuffer at start of cycle and accumulate data until we
transmit at end of cycle.  Accordingly, there is no longer any bneed
for users to choose between jittery-but-correct and buffered-but-laggy 
reports. Accordingly, the daemon 'J' command is no longer necessary.
All the client side options and commands connected with it can go away too.