Change the driver event set to be more orthogonal.
This had the expected effect, the N command can now works on SiRFs
again because the subdevice probe is not called after every driver
switch.

One of the Trimbles has minor regression breakage on this change
(two fewer reports emitted just after packet lock) but I'm not
worrying about that as the TSIP driver needs to be upgraded to
do CYCLE_END, anyway.