Håkan Johansson writes:
I believe the bug report 15348 and fix to be correct.  What one may append
to the discussion is that the variable 'valid' is dual-use:

- make sure that the same measurement is not used twice, ensured as ntpd
  clears it after each use.

- make sure that ntpd cannot read values while gpsd is making an update.
  (this is new)

A sequence of events which realistically (but with low probability) could
trigger such a thing:

- NTPD makes a full read, sets valid = 0.
  (shortly thereafter:)
- GPSD writes a new value, finishing by setting valid = 1
  (almost 1 s passes...)
  (then by chance, GPSD makes it before NTPD:)
- GPSD starts to write values, increases count, valid is still 1
  (during the write operation, the write is interrupted:)
- NTPD makes a full read.  it can, as valid is 1 and count does not change
  sets valid = 0.
- GPSD gets back, and finishes the write, increases count, sets valid to 1

and from this point, it can actually continue, goto 'almost 1 s passes'.
ntpd could thus get a continous stream of bad, but correlated, data.