Get current second from the last GPS timestamp.
ntpshm_pps was guessing the current second for the PPS from the system
clock, yet we have zero trust in the system clock.  So instead grab
the current second from the last GPS serial timestamp.

A radical change for the better, but someone will surely find improvements
with experience on the code.