cast times to unsigned long before printing with %lu.
I was getting warnings about integer type, long format arg, and POSIX
says only that time_t is integral or floating (!!).  It seems it's
always integer, long seems big enough, and the code already assumed
unsigned.

(in test code and clearly a bug, so committing even though we're in a
semi-freeze)