Generate jsongen.py and maskaudit.py from templates...
...so that @PYTHON@ can be substituted with the discovered path to
python.  Avoids the nonportable assumption that "/usr/bin/env python"
will find python, because on pkgsrc systems python2.5 etc. are
installed with no bare python to avoid programs binding to a version
(and hence compiled extensions) that might change.

This should cause no functional change on systems with 'python', and
fix the build on pkgsrc-using systems.