This commit resolves the time overflows in the rgm3800 and mkt3301.
These were causing port problems on systems with 32-bit time.  It turns
out the problem was with the assumption that these devices always
deliver a valid time in $GPRMC.  They don't when the navigation
warning bit (second field 'V') is on!  The NMEA driver now knows.

Also, the code now contains a sanity check - it will log a complaint
if it sees a date moere than a year in the future.  This invariably 
indicates some driver-level problem with time extraction.