Refactor gpspipe to use gps_stream() and gps_open().
So it becomes wire-protocol-independent (no longer generating its own
?WATCH command).  It still reads froom the session socket, rather than
doing gps_poll(), in order to avoide decode overhead.

Building on this, use the new WATCH_DEVICE flag to make gpspipe honor
the device part of a standard source specification, if that is
present.

Make -D an option for enabling client-library progress debugging.

Rename the -f (output-file) option to -o.