>From David Ludlow: replace read(2)/write(2) on sockets with send(2)/recv(2).
He notes:

http://www.opengroup.org/onlinepubs/009695399/functions/recv.html
The recv() function is equivalent to recvfrom() with a zero address_len
argument, and to read() if no flags are used.

http://www.opengroup.org/onlinepubs/009695399/functions/send.html
The send() function is equivalent to sendto() with a null pointer
dest_len argument, and to write() if no flags are used.

write/read vs send/recv is a semantic improvement that just happens to
be necessary for Win32.  It's basically the same as int vs socket_t is.
It's prettier now and avoids an #ifdef when/if Win32 support is added.