Regenerate regression tests after changing to the Vincenty formula.
Rather than computing the value the WGS84 ellipsoid semi-minor axis in
every call to ecef_to_wgs84fix, and having this computation in a few
different source files, use a predefined value. This slightly changes
the output value of ecef_to_wgs84fix - in the least-significant digit,
or about 0.1mm.