Fix the select()-spin bug on disconnected devices.
Addresses Berlios tracker bug #14638: 100% cpu when Bluetooth gps device
vanishes.  Also a couple of reports on the mailing lists.

The underlying problem here was that:

* Disconnecting a USB device causes reads from it to begin returning 0,
  in effect an end-of-file condition.

* select(2) sets an active bit not on "I/O is ready" but on "read would not
  block" -- including the end-of-file condition.

* Therefore, select() will spin any time its fd set includes a disconnected
  device.

The fix is to check for a zero-length read explicitly and always take the
device out of the active set when that happens.

We were actually doing that here, but the test was defective in two ways:

1. The check for a zero return from gpsd_poll(), indicating I/O error
   or zero-length read, needed to be *before* the check for full paccket
   rather than after. This effectively disabled it.

2. There was a conditional arm in the gpsd_poll() code that made it ignore
   zero-length reads for up to a full cycle.  I think this may have been a
   fossil from long ago when I experimented with non-blocking reads on
   devices.

The fix for the gpsmon spin bug is probably quite similar.  That's up next.