Fix the broken handling of our two compass device types in the NMEA driver.
It could never have worked right because there was no mask flag to tell a 
caller that the compass data was valid. TNT has been broken for years; 
why we never got any complaints about the Oceanserver is more of a mystery.

Since it was all broken anyway I renamed and cleaned up that substructure
in the union.  Nothing fills in the "yaw" slots yet with anything but NaN, 
but it's got pitch and roll, so... 

All regression tests pass.