Clear DOPs only when we get a skyview report.
This solves the disappearing epx/epy bug on SiRFs, but it was actually a
systemic problem that affected several drivers.  Navigation solution messages
were clearing DOPs, making it impossible for the error modeller to compute
estimates. New logic: Clear DOPs only when we get a skyview report.  They'll
be regenerated by our visibility-matrix calculation when the skyview sentence
is analyzed.

If a sentence from the device supplies a DOP between the time the visibility
calculation is done and when the next fix is reported, it will override
our computed value.  This might change later!

This change required a regression-test rebuild.