Use 127.0.0.1 for loopback tests.
Previously, the tests use 127.0.1.255.  While it isn't clear if
standards require communications to this address to work, historical
behavior on BSD dating back to 4.4BSD is that 127.0.0.1 is the sole
valid loopback address and the rest of 127/8 is unreachable due to a
blackhole route.  Therefore, avoid addresses other than 127.0.0.1 in
127/8.  Causes et-332 test to pass on NetBSD.