Enhance the libQgpsmm build process.
We now pass libdir and includedir from autotools to qmake to ensure the
library is installed in the same location.
This should avoid an issue on some distributions where 64bit libraries
are living in a different directory. There should be a better way to let
qmake figure this out, but for now this should work.