Pass MAKE variable to setup.py.
When run as "make install" on NetBSD, all is well.  "gmake install"
fails not being able to make target "w".  Passing MAKE=$(MAKE) to
setup.py resolves this, because the child make from setup.py (why is
that happening) is than gmake.  I don't know why calling setup.py from
gmake and then calling (bsd)make from setup.py is a problem, but it
seems to be asking for trouble to switch make implementations in
mid-install.