Bug fixes and improvements for the NTRIP code.
The first thing I had to do to make RTCM work at all, was to remove
the separate poll for the socket (the change in gpsd.c). The same
stuff is done in consume_packet, so there is no loss here. In fact the
duplicated read caused constant lock losses on the RTCM stream because
of missing data, which was already read by the now removed read.

Add RTCM2 passthrough to the UBX driver: versions of the firmware since 7.0
can handle this.

The change in net_ntrip.c adds another string to the valid strings for
rtcm2 to be recognized. See:
http://www.sapos-ni-ntrip.de:2101/sourcetable.htm for the sourcetable
of the server. The mountpoint I am using is EPS_NI.  The problem is
the RTCM1_ data format. The people from sapos confirmed, that this is
a RTCM2 stream and so far it works.

All regression tests pass.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>