Proposed improvement of gpsprof
I used the "gpsprof" tool in conjunction with "gpsfake" to produce
some great looking scatter plots. While working with log files
including thousands of samples I recognized that the result is not
looking good. The CEP circle is covered and the plot looks quite
crowded.

I propose a simple patch to gpsprof which adds three things:
 - Use "dots" instead of "points" when there are more than 1000 samples
 - Plot the CEP circle after the sample points, not before
 - Also plot 95 and 99% plot

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>