The -N option no longer drops privileges.
The price is that only devices specified on the gpsd commmand line will
do PPS, because the calls to set line discipline stop being available
after we drop privileges.  In practice, this shouldn't be a problem,
as PPS devices ar RS232 ports that can't be hotplugged and thus *must*
be specified on the command line.

All regression tests pass.