Make the sock_export=no build.
The way I fixed this extends the library API so some functions which were
previously undefined for shm transport are now defined.  This doesn't change
the binary API of the library in the normal (sock_export=yes) case at all,
so I'm not bumping its version.

No changes in the daemon.  All regression tests pass (in the normal
sock_export=yes build).