Remove proposal that test clients do PCMCIA-like beep codes on status changes.
I had this described as an easy first project, but it turned out to be
nothing of the sort.  The short version is that generating beeps is
surprisingly hard to do in a way that is robust and portable. Support
for beep() is flaky in xterm, and even when it is supported there's no
way to do beeps of varying pitch. X-based clients would have to drag
in a boatload of fragile sound support for this one operation. The
result would be a support mess.  If third-party clients want this,
they can code it themselves.