Removed the proposal about timezone setting.
After meditating on the matter, I've concluded that it's not our job
to know about timezones or calendrical nuances. That would be serious
mission creep, and is best left to domain experts in those areas. If
they  want to write a client that takes UTC and lat/lon and does the
right magic, good.  Our job ends with delivering those inputs.