Fix rpath handling to not leave a security hole.
First, only edit the RPATH to include $CWD and support running tests with the
local copies of the shared libraries *before* installation if we have chrpath
available to remove that tweak and installation time.

Second, simply replace the tweaked RPATH to holdd the configured libdir.
Not a lot of point to trying to edit the path to remove jusat the one
segmennt, because the default RPATH is empty and we know exactly how we
have tweaked it.