In SiRF, don't set the fix timestamp on a MID4.
Longer collection runs show that this often results in bad timestamps
from skyviews not good enough for a fix. There was a comment here
explaining that this was a way to pass a timestamp to NTP, but that
isn't necessary; if there's a fix, NTP will get notified by the MID2
sentence in each cycle.