Attempted fix workaround for the TIOCMIWAIT-hang bug.
The PPS thread hangs whenever you changew baud (or possibly other serial
params) on the devices, so spawning it needs to be deferred until after
autobaud and sync lock have been achieved.

The call to terminatec the thread on device deactivation han't been
put in yet.Pushing this version as a proof of concept so Gary can
test it.