Do not assume little_endian just because __BIG_ENDIAN__ is not defined.
__BIG_ENDIAN__ and __LITTLE_ENDIAN__ are gcc internal defines.
Unfortunately they are not available on all gcc versions and platforms.
So try to use endian.h if neither __BIG_ENDIAN__ nor __LITTLE_ENDIAN__
is defined - and throw an error if we can't retrieve the platform's
endianess from there. endian.h is not available on all platforms,
so we might want to stick with both methods.

This fixes the rtcm2/naujoks regressions at least on s390.