gpsd_hexdump: Always dump in hex mode
This function doesn't always dump the data in hex form.
The problem is that some users on it to always do hex dumps.
For example, the AIS json code use it to dump undecoded data. Depending on the
data it will have either text dump or hex dump, which breaks the json
idempotency.

* This patch doesn't bring any regression. *
scons check and splint passed