Do *not* clear DOPs every time a satellite picture comes from the receiver.
Required a regression-test rebuild.

The immediate reason for this was Savannah bug bug #37810:
satellites_used always zero via gpsd socket with multi nmea GSA/GSV. As the
user reporting said:

The "satellites_used" field in a "struct gps_data_t" filled in by
"gps_read" is always returning zero.

This module emits GNGSA messages in a group of three. My information
is that the first GNGSA pertains to GPS, second to GLONASS, third to
QZSS. It also emits GSV messages using talker id's GL,GP,QZ.

The larger point is that DOPs are likely to be valid for longer than
a GSV reporting cycle; they change only slowly as the actual sat
configuration does.  So it makes sense to retain them.