gpspipe -u time-stamps with sub-second resolution
For simple logging of NMEA data or system performance analysis
it is often desirable to have high resolution time-stamps.
This implementation affects only the client-side, namely gpspipe.
Option -u was added to enable usec-resolution time output.

Furthermore, the default time format has been changed because
it is more desirable to have a standardized and sortable time-tag
default format instead of a locale-dependent one.
Also, the new option -u appends decimal digits to the seconds
which does not make sense for the some %c formats.

A future desirable feature might be to have usec resolution
timestamps on the gpsd protocol messages.

	modified:   gpspipe.c
	modified:   gpspipe.xml

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>