Fix coveraging mode in SConstruct
This patch fixes the coveraging mode, unfortunately the built of the python
bindings requires a dirty hack as it doesn't pickup the global gcc flags.

Right now, it is possible to get coverage data for some of the testsuites,
some fail because the test programs segfault.

Passed testsuites:
 time-regress
 unpack-regress
 json-regress
 maidenhead-locator-regress
 gps-regress
Failed testsuites (segfault):
 rtcm-regress
 aivdm-regress
 packet-regress
 geoid-regress

I still don't know if the segfault is specific to my distro or not
(OpenSUSE 13.1-M4/64bits).
Any feedback welcome!

Signed-off-by: Christian Gagneraud <chgans@gna.org>