Extract ITU-R 1371-4 model and serial fields from AIS Type24 messages.
In ITU-R 1371-4, there are new model and serial fields carved out of
the right-hand end of vendorid, which is reduced from 7 chars to 3.
It is not clear in which minor revision this change took place.  To
cope with older AIS implementations, unpack the trailing bits *both*
ways; truly revision-4-conformant implementations will have up to four
characters of trailing garbage on the vendorid, and older
implementations will have garbafe in the model and serial fields.

This commit also reverts the change to copy only 20 characters rather
than 21 when analyzing the name field in a type21.  There's some mystery
about why this didn't break the rgression tests sooner.