Increase the resilience of the packet getter.
Fail hunt only if we get a second consecutive bad packet and the lexer
is in ground state.  We don't want to fail on a first bad packet
because the source might have a burst of leading garbage after open.
We don't want to fail if the lexer is not in ground state, because
that means the read might have picked up a valid partial packet -
better to go back around the loop and pick up more data.

This change also improves behavior if a read happens to end in mid-packet.

All regression tests pass.