Observer-driver reversion logic.  This is a simple but major change.
Device drivers are now sorted into two categories: observers and controllers.

An "observer" is a device type for which there are no control methods
(that is, no mode switcher, no speed switcher, and no rate switcher).
Note that observers *are* allowed to require initialization sends (such
as wakeup probes or commands to set a sentence mix). Important observers
include Generic NMEA, AIVDM, RTCM2, and RTCM3.

Any other driver is a controller.

The new behavior is simple: each time after you have recognized a
packet to be handled by an observer driver, *revert the session's
device-driver pointer to the previous controller* (if we've seen one).
This means that seeing AIDVM or RTCM packets in an NMEA stream no
longer causes us to lose information about the type of the driver if
it had control methods we might want to use later.

All regression tests pass.