Reinsert the language about known latency in Sratum-00>Stratun-1 connections.
The End Run Technologies page is pretty clear about this.  It says:
"The basic definition of a stratum-1 time server is that it be
directly linked (not over a network path) to a reliable source of UTC
time such as GPS, WWV, or CDMA transmissions"

Besides, it has to be that way or the whole system architecture
wouldn't make sense.