PPS on a pty doesn't work.
This had been breaking the regression tests for a while (since
"Unconditionally enable PPS at startup." to be exact) but I didn't
realize it because an unrelated bug in the build system was usually
turning PPS off for me. Live tests were unaffected.

All regression tests pass with pps either on or off.  Live PPS works.