Regretfully, give up on the raw-regress test.
It wasn't part of the normal test dequence for good reason.  There are
just too many silly ways for it to go wrong, failing or even hanging.
due to bad interactions between logfiles with trailing content that isn't
a whole packet and various quirks of the test framework.

Given that even raw mode is a fossil - let alone superraw mode - there's
not enough value in this test to be worth the hassle.