RTCM V2: use scons to determine the location of <endian.h>.
There are three cases to consider:

- The compiler has built-in endianness macros. In this case, endian.h is
  not needed at all. The predefined symols are called
  __ORDER_BIG_ENDIAN_, __ORDER_LITTLE_ENDIAN__ and __BYTE_ORDER__.

- <endian.h> exists.

- <sys/endian.h> exists.

This commit covers all three cases at configure time, as Greg Troxel
wanted it.

Relevant discussion can be found on the GCC mailing list:
http://gcc.gnu.org/ml/gcc-help/2007-07/msg00342.html

Of course, the optimal solution would get rid of the bit field
structures and use bits.h to extract the values from the buffer in an
endianess-independent way.

Signed-off-by: Beat Bolli <bbolli@ewanet.ch>
Signed-off-by: Eric S. Raymond <esr@thyrsus.com>