Remove ad-hoc static library libgps.a build.
It wasn't used by any top-level build target, but for some strange
reason scons was promoting the target so it was always built. (This
could be a "feature" of the StaticLibrary() primitive, I suppose.)
This could cause problems with parallel builds; while the linker
apparently prefers a shared library to its static equvalent when
satisying an -l flag, unexpected static linking can happen in a
parallel build if libgps.a happens to get built before libgps.*.so.