libgpsmm: add is_open method
The gpsmm constructor effectively hides the useful status flag returned
by the underlying gps_open call.

It is extremely useful for client code to be able to check that the
gpsmm constructor was successful. To this end, add a gpsmm::is_open method
which simply checks that the gpsmm::to_user member is non-NULL. A NULL
gpsmm::to_user member can only mean that the constructor was not successful.

Signed-off-by: Simon Dawson <spdawson@gmail.com>
Signed-off-by: Eric S. Raymond <esr@thyrsus.com>