gpsd: add TOS and ECN support
This attempts to enable marking gpsd packets to use a lowdelay
TOS marking in the packet headers. It also enables the ECN bit,
to lower dropping probability for systems that respect ECN.

It doesn't actually DO anything with the ECN bit as at a rate
of 1 per second, there isn't any rate reduction to be had.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>