Fix paths to gpsd.hotplug in udev rules
The gpsd build system allows us to specify where to install udev rules
with "scons udevdir=...", but the rules file itself will always contain
(hardcoded) paths to /lib/udev. Which makes it broken.

This patch makes it so that the references to gpsd.hotplug stays in sync
with where the rules file is installed, by adding a @udevdir@ "magic
marker" in the rules file (now renamed to gpsd.rules.in) and then
substituting that for the real udevdir install path at build time.

Signed-off-by: Bjørn Forsman <bjorn.forsman@gmail.com>
Signed-off-by: Eric S. Raymond <esr@thyrsus.com>