Fix PPS with large offsets.
This was broken by commit 575444. The check if the PPS time isn't
referenced with an old message time used PPS offset instead of the
difference between local PPS time and local time of the last message.
The PPS sample was reported via SHM or chrony socket only if the offset
was between -1 and 1000001 seconds.

Keep the local timestamp of the last fix in struct gps_device_t in
addition to the real timestamp and use it to check the PPS delay.
Rename the l_offset variable to delay to avoid further confusion.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>