Reimplement Zodiac control-send safely for word-oriented architectures.
This was a bug revealed by Debian porterbox testing.  We used to use
a cast from (char *) to (unsigned short *) that was problematic on
word-oriented architectures like the PowerPC.  Instead, memcpy the
message data to a word array, which the compiler will create with
correct alignment.

All regression tests, cppcheck, and splint pass.